legacy/original_src/requirements.txt
fastmcp
requests
python-dotenv
openai-agents
fastapi
uvicorn[standard]
websockets
openai
mysql-connector-python
